# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountLeaseHandle": {
        "additionalProperties": false,
        "description": "不包含 Account ID、登录名或凭证引用的 Worker 安全投影。",
        "properties": {
          "accountHandle": {
            "pattern": "^ah_[A-Za-z0-9_-]{16,128}$",
            "title": "Accounthandle",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "fencingToken": {
            "minimum": 1.0,
            "title": "Fencingtoken",
            "type": "integer"
          },
          "heartbeatAfterSeconds": {
            "minimum": 60.0,
            "title": "Heartbeatafterseconds",
            "type": "integer"
          },
          "leaseId": {
            "format": "uuid",
            "title": "Leaseid",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AccountLeaseStatus"
          }
        },
        "required": [
          "leaseId",
          "accountHandle",
          "fencingToken",
          "status",
          "heartbeatAfterSeconds",
          "expiresAt"
        ],
        "title": "AccountLeaseHandle",
        "type": "object"
      },
      "AccountLeaseStatus": {
        "description": "账号租约的持久化状态。",
        "enum": [
          "ACTIVE",
          "RELEASED",
          "EXPIRED",
          "REVOKED"
        ],
        "title": "AccountLeaseStatus",
        "type": "string"
      },
...
      "ErrorCode": {
        "description": "可以安全暴露给调用方的稳定错误码。",
        "enum": [
          "INVALID_REQUEST",
          "VALIDATION_FAILED",
          "AUTHENTICATION_REQUIRED",
          "AUTHENTICATION_FAILED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "PRECONDITION_FAILED",
          "POOL_EXHAUSTED",
          "CONSTRAINT_UNSATISFIED",
          "LEASE_EXPIRED",
          "LEASE_FENCED",
          "CREDENTIAL_EXPIRED",
          "ORIGIN_NOT_ALLOWED",
          "SECRET_GRANT_EXPIRED",
          "SECRET_GRANT_REPLAYED",
          "SECRET_GRANT_REVOKED",
          "SESSION_CREATION_IN_PROGRESS",
          "SESSION_UNAVAILABLE",
          "SESSION_INTEGRITY_FAILED",
          "PROVIDER_UNAVAILABLE",
          "DEPENDENCY_UNAVAILABLE",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "FieldViolation": {
        "additionalProperties": false,
        "description": "请求字段的单个校验问题。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "code"
        ],
        "title": "FieldViolation",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "Atlas 扩展的 RFC Problem Details。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errorCode": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            },
            "title": "Violations",
            "type": "array"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "errorCode",
          "requestId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Atlas TestOps Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/account-leases/{leaseId}": {
      "get": {
        "description": "只返回当前租约控制字段和不透明 Account Handle。",
        "operationId": "get_account_lease_internal_v1_account_leases__leaseId__get",
        "parameters": [
          {
            "in": "path",
            "name": "leaseId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Leaseid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Tenant-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Atlas-Actor-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountLeaseHandle"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "租约请求或 Execution Deadline 无效"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "缺少可信 Worker 或 Platform 身份"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "运行角色或生产环境策略拒绝请求"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "资源不存在或不可见"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "账号池耗尽、Lease 过期或被 Fencing"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "角色、标签、能力或认证要求无法满足"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "服务内部错误"
          }
        },
        "summary": "读取账号租约",
        "tags": [
          "identity-runtime"
        ]
      }
    },
...
  }
}